requirements.txt
pytest==9.0.2
pytest-cov==5.0.0
pytest-html==4.1.1
pyserial==3.5